+++ src/aks-preview/azext_aks_preview/tests/latest/test_validators.py
@@ -124,6 +124,9 @@ class DisableWindowsOutboundNatNamespace:
         self.os_type = os_type
         self.disable_windows_outbound_nat = disable_windows_outbound_nat
 
+class UpdateOutBoundTypeNamespace:
+    def __init__(self, outbound_type):
+        self.outbound_type = outbound_type
 
 class TestMaxSurge(unittest.TestCase):
     def test_valid_cases(self):
@@ -232,6 +235,17 @@ class TestDisableWindowsOutboundNAT(unittest.TestCase):
             validators.validate_disable_windows_outbound_nat(DisableWindowsOutboundNatNamespace("invalid", True))
         self.assertTrue('--disable-windows-outbound-nat can only be set for Windows nodepools' in str(cm.exception), msg=str(cm.exception))
 
+class TestUpdateOutBoundType(unittest.TestCase):
+    def test_pass_if_os_type_loadbalancer(self):
+        validators.validate_outbound_type_in_update(UpdateOutBoundTypeNamespace("loadBalancer"))
+
+    def test_pass_if_os_type_managedNATGateway(self):
+        validators.validate_outbound_type_in_update(UpdateOutBoundTypeNamespace("managedNATGateway"))
+
+    def test_pass_if_os_type_UDR(self):
+        with self.assertRaises(CLIError) as cm:
+            validators.validate_outbound_type_in_update(UpdateOutBoundTypeNamespace("userDefinedRoute"))
+        self.assertTrue('Invalid outbound type userDefinedRoute, supported values are loadBalancer and managedNATGateway' in str(cm.exception), msg=str(cm.exception))
 
 class ValidateAddonsNamespace:
     def __init__(self, addons):
@@ -598,6 +612,46 @@ class TestValidateApplicationSecurityGroups(unittest.TestCase):
         validators.validate_application_security_groups(
             namespace
         )
+
+class MaintenanceWindowNameSpace:
+    def __init__(self, utc_offset=None, start_date=None, start_time=None):
+        self.utc_offset = utc_offset
+        self.start_date = start_date
+        self.start_time = start_time
+
+class TestValidateMaintenanceWindow(unittest.TestCase):
+    def test_invalid_utc_offset(self):        
+        namespace = MaintenanceWindowNameSpace(utc_offset="5:00")
+        err = '--utc-offset must be in format: "+/-HH:mm". For example, "+05:30" and "-12:00".'
+        with self.assertRaises(InvalidArgumentValueError) as cm:
+            validators.validate_utc_offset(namespace)
+        self.assertEqual(str(cm.exception), err)
+    
+    def test_valid_utc_offset(self):        
+        namespace = MaintenanceWindowNameSpace(utc_offset="+05:00")
+        validators.validate_utc_offset(namespace)
+
+    def test_invalid_start_date(self):        
+        namespace = MaintenanceWindowNameSpace(start_date="2023/01/01")
+        err = '--start-date must be in format: "yyyy-MM-dd". For example, "2023-01-01".'
+        with self.assertRaises(InvalidArgumentValueError) as cm:
+            validators.validate_start_date(namespace)
+        self.assertEqual(str(cm.exception), err)
+    
+    def test_valid_start_datet(self):        
+        namespace = MaintenanceWindowNameSpace(start_date="2023-01-01")
+        validators.validate_start_date(namespace)
+    
+    def test_invalid_start_time(self):        
+        namespace = MaintenanceWindowNameSpace(start_time="3am")
+        err = '--start-time must be in format "HH:mm". For example, "09:30" and "17:00".'
+        with self.assertRaises(InvalidArgumentValueError) as cm:
+            validators.validate_start_time(namespace)
+        self.assertEqual(str(cm.exception), err)
+    
+    def test_valid_start_time(self):        
+        namespace = MaintenanceWindowNameSpace(start_date="00:30")
+        validators.validate_start_time(namespace)
 
 
 if __name__ == "__main__":